% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{cetesb}
\alias{cetesb}
\title{Base de dados da Cetesb}
\format{
\describe{
  \item{estacao_cetesb}{estacao_cetesb}
  \item{poluente}{poluente}
  \item{data}{data}
  \item{hora}{hora}
  \item{concentracao}{concentracao}
  \item{lat}{lat}
  \item{long}{long}
}
}
\usage{
cetesb
}
\description{
Base de dados da Cetesb
}
\keyword{datasets}
